+++ stellar-sdk/src/commonMain/kotlin/com/soneso/stellar/sdk/xdr/StellarValueTypeXdr.kt
@@ -8,12 +8,14 @@ package com.soneso.stellar.sdk.xdr
  * enum StellarValueType
  * {
  *     STELLAR_VALUE_BASIC = 0,
- *     STELLAR_VALUE_SIGNED = 1
+ *     STELLAR_VALUE_SIGNED = 1,
+ *     STELLAR_VALUE_EMPTY_TX_SET = 2
  * };
  */
 enum class StellarValueTypeXdr(val value: Int) {
   STELLAR_VALUE_BASIC(0),
-  STELLAR_VALUE_SIGNED(1);
+  STELLAR_VALUE_SIGNED(1),
+  STELLAR_VALUE_EMPTY_TX_SET(2);
 
   companion object {